requirements.txt
python-dotenv

requests 
playwright
beautifulsoup4

langchain
langchain-openai